amend: 2022-001 to protect from rights being infringed upon
This addition makes it the right of a Farer member to not
infringe upon the rights of others or knowingly obstruct
others from exercising their obligations or infringing
upon their rights. It also explicitly guarntees the right
to not have the infringement or obstruction occur to
themself, as a Farer member.